Reconstruct: Now implemented as lazy evaluation.
The implementation is working left most first.
Each hyper edge is read and then 3 steps of calculation are made:
edge step to find new trees from the new edge
advance step to update if we found a new identifier
develop step to create all trees that can be created from the new trees and edge.

Signed-off-by: Eytan Singher <eytan.singher@gmail.com>