Bug Fix: Fixed existential support and sample.
Added debug logging for each word and rule being worked on.
Fixed shift patterns to take care of the case that one of the vars is
equal to one of the patterns. Removed prints from tactic as the were
buggy and resulted in a really long run. Fixed encoding so vars and main
term will both stay in transformation to patterns. Existential sample
now uses len to make xs different then placeholder(0). Removed xs from
vars in existential rules until we have a tag system with types.